Restructured the code to different classes, login, usercreation and the navigation methods are in their own files now :)
